requirements.txt
Django~=4.2.2
django-ninja
django-environ
psycopg2
psycopg2-binary
pytest-django